Exposed several additional standup parameters
This PR increases the list of parameters - employed by both
"standalone" and "deployer" methods - to increase the ability to
properly express more feature rich scenarios.

Main motivation was to provide the ability to capture the scenarios
described at
https://github.com/tlrmchlsmth/prefill-decode-experiments/tree/main,
using the usual llm-d-benchmark mechanism (environment variables defined
in a "scenario" file).

Signed-off-by: maugustosilva <maugusto.silva@gmail.com>