[Setup] feat: first implementation of "parameter sweep" for standup.
This PR builds on #184, by adding support for "parameter sweeps" on
setup/standup. The "experiment file" (`yaml`) was expanded to include a
`setup:` and `run:` sections. On the former, deployment parameters, such
as the number of `prefill` and `decode` `pods` can be specified.

Evidently, the "parameter sweep" for standup only makes sense when
invoked with `e2e.sh`, and the implementation reflects it accordingly.

Additionally, I have started a "consolidation" of scenarios.

Signed-off-by: maugustosilva <maugusto.silva@gmail.com>