[Setup] Extensive re-organization of scenarios
Created a new directory, `guides` (with a symbolic link, `well-lit`) to
indicate how can a user deploy these specific scenarios.
Also added a simple, minimalistic `examples` for different accelerators.
Finally, added a significant amount of documentation, in the form of
comments, directly on the files.

Fix a bug on step 5, which impacted the user's ability to execute
`run.sh`

Signed-off-by: maugustosilva <maugusto.silva@gmail.com>